+++ src/core/Application.h
@@ -95,7 +95,7 @@ public slots:
 	void triggerAction(int identifier, const QVariantMap &parameters = {}, ActionsManager::TriggerType trigger = ActionsManager::UnknownTrigger) override;
 
 protected:
-	void scheduleUpdateCheck(int interval);
+	void scheduleUpdateCheck(quint64 interval);
 	static void setLocale(const QString &locale);
 
 protected slots:
